requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
boto3==1.28.62
botocore==1.31.62
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
fastapi==0.103.2
ffmpeg-python==0.2.0
future==0.18.3
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-speech==2.21.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
httptools==0.6.0
idna==3.4
jmespath==1.0.1
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rq==1.15.1
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
numpy~=1.26.0
keras~=2.14.0
scikit-learn~=1.3.1
nltk~=3.8.1
dill~=0.3.7
lime~=0.2.0.1
moviepy~=1.0.3